requirements.txt
openai>=1.30.0
rich>=13.7.0
python-dotenv>=1.0.1
